p9: Make wstat changes argument its own type.
