+++ p9.go
@@ -2,7 +2,6 @@ package p9
 
 import (
 	"math"
-	"time"
 )
 
 const (
@@ -72,79 +71,4 @@ func (t QIDType) encode(e *encoder) {
 
 func (t *QIDType) decode(d *decoder) {
 	d.Decode((*uint8)(t))
-}
-
-// Stat is a stat value.
-type Stat struct {
-	Type   uint16
-	Dev    uint32
-	QID    QID
-	Mode   uint32 // TODO: Make a Mode type?
-	ATime  time.Time
-	MTime  time.Time
-	Length uint64
-	Name   string
-	UID    string
-	GID    string
-	MUID   string
-}
-
-func (s Stat) dirEntry() DirEntry {
-	return DirEntry{
-		Type:   s.QID.Type,
-		Mode:   s.Mode,
-		ATime:  s.ATime,
-		MTime:  s.MTime,
-		Length: s.Length,
-		Name:   s.Name,
-		UID:    s.UID,
-		GID:    s.GID,
-		MUID:   s.MUID,
-	}
-}
-
-func (s Stat) size() uint16 {
-	return uint16(47 + len(s.Name) + len(s.UID) + len(s.GID) + len(s.MUID))
-}
-
-func (s Stat) encode(e *encoder) {
-	e.Encode(s.size())
-	e.Encode(s.Type)
-	e.Encode(s.Dev)
-	e.Encode(s.QID)
-	e.Encode(s.Mode)
-	e.Encode(s.ATime)
-	e.Encode(s.MTime)
-	e.Encode(s.Length)
-	e.Encode(s.Name)
-	e.Encode(s.UID)
-	e.Encode(s.GID)
-	e.Encode(s.MUID)
-}
-
-func (s *Stat) decode(d *decoder) {
-	var size uint16
-	d.Decode(&size)
-
-	r := d.r
-	d.r = &LimitedReader{
-		R: r,
-		N: uint32(size),
-		E: ErrLargeStat,
-	}
-	defer func() {
-		d.r = r
-	}()
-
-	d.Decode(&s.Type)
-	d.Decode(&s.Dev)
-	d.Decode(&s.QID)
-	d.Decode(&s.Mode)
-	d.Decode(&s.ATime)
-	d.Decode(&s.MTime)
-	d.Decode(&s.Length)
-	d.Decode(&s.Name)
-	d.Decode(&s.UID)
-	d.Decode(&s.GID)
-	d.Decode(&s.MUID)
 }
